Added scheduled methods to update BookingTimes in DB and changed name on AllTimeSlots-view
Added scheduled method that once per day, every day, adds booking times for the day two weeks from now to the table BookingTimes in the db. Added scheduled method that once per day, every day, removes the old booking times of yesterday. Changed name of the view AllTimeSlots to BookingsView. Changed the insert in DBController so that it operates on the BookingsView instead.

Co-Authored-By: ArvidHolmqvist <54509607+ArvidHolmqvist@users.noreply.github.com>