Added checks to AdditionalInfo and fixed next & previous buttons
AdditionalInfo now validates the forms for name, email and phone number. A user cant go to the next page unless all three are valid.

Also added default values to the guest, date and timelist page. If a user selects a number of guests then goes to the next page and then back, the same number of guests is still selected and the 'next' button is not disabled